# docs/swagger.yaml
definitions:
...
  github_com_michaelwp_student_attendance_internal_models.Student:
    properties:
      classes_id:
        type: integer
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phone:
        type: string
      student_id:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /students:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all students
      parameters:
      - default: 10
        description: Number of students to return (max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of students to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of students retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all students
      tags:
      - Students
    post:
      consumes:
      - application/json
      description: Create a new student in the system
      parameters:
      - description: Student data
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/github_com_michaelwp_student_attendance_internal_models.Student'
      produces:
      - application/json
      responses:
        "201":
          description: Student created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new student
      tags:
      - Students
  /students/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a student from the system
      parameters:
      - description: Student database ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Student deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid student ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete student
      tags:
      - Students
    get:
      consumes:
      - application/json
      description: Retrieve a specific student by their database ID
      parameters:
      - description: Student database ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Student retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid student ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Student not found
          schema:
            additionalProperties: true
            type: object
      summary: Get student by ID
      tags:
      - Students
    put:
      consumes:
      - application/json
      description: Update a student's information
      parameters:
      - description: Student database ID
        in: path
        name: id
        required: true
        type: integer
      - description: Student data
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/github_com_michaelwp_student_attendance_internal_models.Student'
      produces:
      - application/json
      responses:
        "200":
          description: Student updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update student
      tags:
      - Students
  /students/class-id/{classId}:
    get:
      consumes:
      - application/json
      description: Retrieve all students in a specific class
      parameters:
      - description: Class ID
        in: path
        name: classId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Students retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid class ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get students by class
      tags:
      - Students
  /students/student-id/{studentId}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific student by their student ID
      parameters:
      - description: Student ID (not database ID)
        in: path
        name: studentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Student retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Student ID is required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Student not found
          schema:
            additionalProperties: true
            type: object
      summary: Get student by student ID
      tags:
      - Students
...
swagger: "2.0"